1.在mongo中存储电影院信息 2.计算给定经纬度距离电影院的距离

+++ common/src/main/java/com/cxy/entry/Cinema.java
@@ -5,7 +5,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -30,12 +29,12 @@ public class Cinema implements Serializable {
     /**
      * 经度
      */
-    private BigDecimal x;
+    private double x;
 
     /**
      * 纬度
      */
-    private BigDecimal y;
+    private double y;
 
     /**
      *
